kernel-shark: Allow specifying PKG_CONFIG_DIR on cmake command line
On some distributions (e.g. NixOS), it is not possible to write files
to directories owned by different packages. For this reason, we cannot
always install libkshark.pc file to the directory reported by
pkg-config.

This commit allows to specify where to install the .pc file on cmake
commandline via -DPKG_CONGIG_DIR=... When specified, automatic
detection of .pc install directory is skipped.

Link: https://lore.kernel.org/linux-trace-devel/20210317164101.30848-2-michal.sojka@cvut.cz

Signed-off-by: Michal Sojka <michal.sojka@cvut.cz>
Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>