kernel-shark: Integrate install_libkshark-devel.sh
So far "install_libkshark-devel.sh" was an unofficial script that
was used to help installing the "libkshark-devel" component. It was
introduced, because the CMake syntax for doing this is quite hard to
memorize. Here we make this script an integrated part of the build
system. The message output is made consistent with the messages
printed by "install_gui.sh". The usage of the script is explained
in README.

Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>