kernel-shark: Fix KS_DEFINE_PLUGIN_CONTEXT macro
The KS_DEFINE_PLUGIN_CONTEXT macro implements methods that are used
to deal with plugin-specific context objects. However, the methods
defined by the macro (as it is now) are static and cannot be used in
plugins, built from multiple source files.

If we want the methods to be declared in a header and used in different
source files, we have to make sure that those methods are not visible
outside of the corresponding plugin. The opposite can be dangerous
in the case when the macro is used in multiple plugins and those plugins
are loaded together. In such a case the symbol resolving will fail,
resulting in undefined behavior.

Link: https://lore.kernel.org/linux-trace-devel/20210428134730.187533-4-y.karadz@gmail.com

Fixing: 15df009 (kernel-shark: Add KS_DEFINE_PLUGIN_CONTEXT macro)
Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>