kernel-shark: Add EventFieldPlot plugin
The plugin allows the user to visualize the recorded value of a given
data field from a given trace event. The core logic that implements
the processing of the data and the visualization itself is rather
simple. It is implemented in:
event_field_plot.h,
event_field_plot.c and
EventFieldPlot.cpp

The plugin also registers its own dialog, that allows the user to
select the event and field to be visualized. The widget of the dialog
gets implemented in:
EventFieldDialog.hpp and
EventFieldDialog.cpp

Link: https://lore.kernel.org/linux-trace-devel/20210429154421.255872-2-y.karadz@gmail.com

Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>