kernel-shark: Add LatencyPlot plugin
The plugin allows the user to visualize the latency between two
events under the condition that the values of given data fields
in the two events are identical (for example having the same PID).
The core logic that implements the processing of the data and
the visualization itself is implemented in:

src/plugins/latency_plot.h,
src/plugins/latency_plot.c and
src/plugins/LatencyPlot.cpp

The plugin also registers its own dialog, that allows the user to
select the events (and the matching field) to be visualized. The
widget of the dialog gets implemented in:

src/plugins/LatencyPlotDialog.hpp
src/plugins/LatencyPlotDialog.cpp

Link: https://lore.kernel.org/linux-trace-devel/20210429154421.255872-3-y.karadz@gmail.com

Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>