kernel-shark: Add check for return of ksmodel_get_cpu_front()
When I loaded two trace.dat files (host and guest), but the mapping of
which host thread is associated to which guest vCPU was missing from the
file, it caused a SEGFAULT. That's because in fillTaskGraph(), the
lamGetPidCPU() calls ksmodule_get_cpu_front() which returns a negative
number and does not set index. But the next line checks data[index] where
index is some random number, and the application crashes.

By checking the return of ksmodule_get_cpu_front(), and if it is negative
do not reference data[index] and just let eFront be nullptr, the
application shows no mapping, but at least it does not crash.

Link: https://lore.kernel.org/linux-trace-devel/20210507093848.72bf0bed@gandalf.local.home

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>