kernel-shark: Treat all tracing headers as library headers
The CMake scripts used to discover all tracing libraries are modified
such that the full path to the header is determined. Later there is
no need to add that part when the headers are included. Using angular
brackets ('<' and '>') denotes that it is a library header and not a
local one.

1. Note that this patch actually fixes a build failure, caused by the
following commit in libtracefs:
880885d (libtracefs: Remove "traceevent/" from referencing libtraceevent headers)
It seems that so far the discovery of the headers was working more or
less by chance.

2. Note that after applying the patch you may need to clean the Cmake
cache (running cmake_clean.sh).

Link: https://lore.kernel.org/linux-trace-devel/20210726094141.191579-2-y.karadz@gmail.com

Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>
Acked-by: Steven Rostedt (VMware) <rostedt@goodmis.org>