kernel-shark: Hide CPUs with no data in them
A hash table, storring all idle CPUs is added to the session context.
If a CPU does not contain any data (is idle), its plot is not shown
by default when KernelShark starts. Note that, if the option '--cpu'
is used, it has priority and if idle CPUs are selected, the
corresponding empty CPU plots will be displayed.

Bugzilla: https://bugzilla.kernel.org/show_bug.cgi?id=215677
Tested-by: Steven Rostedt (Google) <rostedt@goodmis.org>
Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>