kernel-shark: Do not truncate multi-line events
Some events have more than one line. The libkshark-tepdata.c get_info
function did a string search for the first occurrence of '\n' and set it
to '\0', with the comment of removing trailing newlines. Unfortunately,
it removed more than the trailing newline and removed most of the event.
This is particularly true with stack traces.

Instead, use the trace_seq.len to check the last written character, and if
that is a newline, remove it.

Fixes: 836ce858246b7 ("kernel-shark: Add stream interface for trace-cmd data")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>
Signed-off-by: Yordan Karadzhov <y.karadz@gmail.com>