kernelshark: Allow to reduce a bit more the graph height
Reduce minimum height of the graph widget which allows the user (using the
splitter widget) to expand a bit more the trace viewer (KsTraceViewer)
which provide the table view (KsTableView).
This is very useful when a user has a very small screen.
With this configuration, at least one CPU core on the graph (KsTraceGraph)
can be seen, so the graph widget is still perfectly usable.

Signed-off-by: Benjamin ROBIN <dev@benjarobin.fr>
Signed-off-by: Yordan Karadzhov <y.karadz@gmail.com>