Documentation: Update latency description for migration disabled
Starting in 5.11, the Linux kernel implemented a "migrate disabled"
counter for each task. When the counter is non-zero, the task will be
temporarily pinned to its CPU. This is useful for accessing per CPU data
that is protected by a per CPU mutex. The task can still be preempted, but
it must not leave the CPU. When its "migrate disabled" counter is greater
than zero, it will not migrate.

The latency field of the tracer now shows this number. Updated the
documentation to explain it.

Reported-by: Jorge Merlino <jorge.merlino@canonical.com>
Closes: https://lore.kernel.org/all/fe35c5eb-7132-4203-9d6f-0a8272b114fa@canonical.com/
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>
Signed-off-by: Yordan Karadzhov <y.karadz@gmail.com>